requirements.txt
eel
pymysql
cryptography
SQLAlchemy
Telethon
requests
Pillow
python-dotenv
pyinstaller
